style: to component - I
