+++ frontend/src/components/component/index.ts
@@ -1,1 +1,7 @@
 export { default as IconItem } from './IconItem'
+
+export { default as MarkdownEditor } from './MarkdownEditor.vue'
+
+export { default as RichTextEditor } from './RichTextEditor.vue'
+
+export { default as WebTerminal } from './WebTerminal.vue'
